Refactor bitbucket to its own class
