+++ Config.js
@@ -2,8 +2,17 @@ require('dotenv').config()
 
 module.exports = {
   bitbucket: {
-    type: 'basic',
-    username: process.env.USERNAME,
-    password: process.env.PASSWORD
+    auth: {
+      type: 'basic',
+      username: process.env.USERNAME,
+      password: process.env.PASSWORD
+    },
+    
+    teamName: process.env.TEAMNAME
+
+  },
+
+  slack: {
+    webhook: process.env.SLACK_WEBHOOK_URL
   }
 }
